Rework MediaDiscoverer
Disable C++11 stuff for now. The code doesn't use it, and I'm not sure
it's a good idea to enforce it if we want a widespread adoption.
Remove references to c_object.
This is unsatisfactory as is, since it enforces using this class as a
pointer to be able to copy it, while we receive our other instances via
reference. We might want to add our own refcounting, or go back to move
semantic, though it's still a usage restriction