MediaTrack: Split 3.x and 4.x implementations.
The 4.x version will store a libvlc_media_track_t instead of all its
fields, which is now possible since the track structure is refcounted.
This will also allow us to provide the original tracks back to any
libvlc function that might require it.
The 3.x version will stick to copying each field to the MediaTrack class
since the original libvlc_media_track_t is released as soon as we're
done constructing a MediaTrack vector